splitting tracers to separate files
